moving components into their respective folders
